requirements.txt
asyncio
async_timeout
git+git://github.com/Rapptz/discord.py.git@6dd8845e4f38f22e226883234d369d43c7629f1a
traceback2
psutil
python-dotenv
yt_dlp